requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-jose[cryptography]>=3.3.0
google-auth>=2.22.0
requests>=2.28.0
pydantic>=2.0.0